Add driver CPU Time to the metrics system
